Improvements to snapshots and yb-admin, and some additional cleanup
Summary:
- When requesting the list of all snapshots via `yb-admin` and the snapshot id is specified, do not crash if the server sends us multiple snapshots.
- Clean up output formatting in`yb-admin`
- Add another function, `PBEnumToString` in the `PB_ENUM_FORMATTERS` macro.  This allows to say `PBEnumToString(protobuf_defined_enum_value)` to get a string representation without having to use a `.._Name` function, and without being confused about what overload of `ToString` might get chosen.
- Automatically try using shared third-party directories in an NFS environment if the default thirdparty directory (`$YB_SRC_ROOT/thirdparty`) has a file `.yb_thirdparty_do_not_use`.
- Import `FindBoost.cmake` from the latest version of CMake.
- Remove `PrecompiledHeader.cmake`, which we don't use.
- Move `yb/rocksdb/util/string_util.h` to `yb/util/string_util.h` -- string utilities should not be specific to RocksDB.
- The `list_tablets` `yb-admin` command now outputs the leader of each tablet (that information is already available in the response). This allows to dramatically reduce the number of remote commands  in the backup script.
- Adding a `--cmake-args` argument to `yb_build.sh`.

Test Plan:
- Jenkins
- Manual testing with `yb-admin`

Reviewers: bharat, oleg, sergei, bogdan

Reviewed By: bogdan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D4055

+++ src/yb/rocksdb/util/autovector_test.cc
@@ -24,16 +24,18 @@
 
 #include "yb/rocksdb/env.h"
 #include "yb/rocksdb/util/autovector.h"
-#include "yb/rocksdb/util/string_util.h"
+#include "yb/util/string_util.h"
 #include "yb/rocksdb/util/testharness.h"
 #include "yb/rocksdb/util/testutil.h"
 
 namespace rocksdb {
 
-using namespace std;
+// TODO(mbautin): if we decide to keep autovector, get rid of "using namespace std".
+// But we'll probably replace it with boost::small_vector.
+using namespace std; // NOLINT
 
 class AutoVectorTest : public testing::Test {};
-const unsigned long kSize = 8;
+const size_t kSize = 8;
 
 namespace {
 template <class T>
